Better exp-integral-e computation and fix for incomplete-gamma-tail.
For exp-integral-e, use the series for small z and the
incomplete-gamma-tail for near the negative real axis.  Otherwise, use
the continued fraction.

In incomplete-gamma-tail, we were using the continued fraction instead
of the incomplete-gamma function for the region just below the negative
real axis.  We should use the cf except in that region.